+++ Homie Hand/src/assets/Grass.java
@@ -11,9 +11,9 @@ public class Grass extends GameObject {
 	// r = 0, g = 102, b = 0
 	private BufferedImage grass;
 	
-	public Grass(int x, int y, int width, int height, int scale, String ID) {
+	public Grass(float x, float y, float width, float height, float scale, String ID) {
 		super(x, y, width, height, scale, ID);
-		grass = ImageLoader.getSpriteSheet(ImageLoader.BLOCK_SPRITESHEET).getSubimage(0, 0, width, height);
+		grass = ImageLoader.getSpriteSheet(ImageLoader.BLOCK_SPRITESHEET).getSubimage(0, 0, (int)width, (int)height);
 	}
 	
 	public void tick(LinkedList<GameObject> gameObjects) {
@@ -21,6 +21,6 @@ public class Grass extends GameObject {
 	}
 	
 	public void render(Graphics g) {
-		g.drawImage(grass, x, y, scaledWidth, scaledHeight, null);
+		g.drawImage(grass, (int)x, (int)y, (int)scaledWidth, (int)scaledHeight, null);
 	}
 }
